pyproject.toml
[project]
name = "mr-robot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "punq>=0.7.0",
    "pydantic-settings>=2.9.1",
    "tinkoff-investments>=0.2.0b111",
    "vectorbt>=0.27.3",
]

[dependency-groups]
dev = [
    "ruff>=0.11.8",
]
